+++ assets/javascripts/kitten/components/form/dropdown-select/dropdown-select.stories.js
@@ -1,8 +1,7 @@
 import React, { useState } from 'react'
-import { text, boolean, select, number } from '@storybook/addon-knobs'
 import { DropdownSelect } from './index'
 import { Button } from '../../../index'
-import { Grid, GridCol } from '../../../components/layout/grid'
+import { Grid, GridCol } from '../../layout/grid'
 import { ArrowIcon } from '../../graphics/icons/arrow-icon'
 import { Text } from '../../atoms/typography/text'
 import { DocsPage } from 'storybook/docs-page'
@@ -45,34 +44,36 @@ const options = [
   },
 ]
 
-export const Default = () => {
+const args = {
+  error: false,
+  valid: false,
+  disabled: false,
+  hideLabel: false,
+  combobox: false,
+  controlled: false,
+  labelText: 'label',
+  size: 'normal',
+  variant: 'andromeda',
+  comboboxButtonLabelText: 'label',
+  noResultText: 'No results',
+  menuZIndex: 10000,
+}
+
+export const Default = props => {
   const [labelProps, setLabelProps] = useState(null)
 
   return (
     <Grid>
       <GridCol offset="1" col="8">
         <DropdownSelect
-          id={text('id', 'dropdown-select')}
-          error={boolean('error', false)}
-          valid={boolean('valid', false)}
-          disabled={boolean('disabled', false)}
-          hideLabel={boolean('hide label?', false)}
-          combobox={boolean('combobox?', false)}
-          labelText={text('LabelText', 'label')}
+          id="dropdown-select"
+          uncontrolled
           options={options}
           labelPropsGetter={passedLabelProps => {
             passedLabelProps && setLabelProps(passedLabelProps())
           }}
-          size={select(
-            'size',
-            ['tiny', 'normal', 'big', 'huge', 'giant'],
-            'normal',
-          )}
-          variant={select('variant', ['andromeda', 'orion'], 'andromeda')}
           defaultSelectedValue="focus"
-          comboboxButtonLabelText={text('Buton aria-label', 'label')}
-          noResultText={text('No results text', 'No results')}
-          menuZIndex={number('menuZIndex', 1000)}
+          {...props}
         />
         <p>
           Integer posuere erat a ante venenatis dapibus posuere velit aliquet.
@@ -94,52 +95,38 @@ export const Default = () => {
   )
 }
 
-export const WithDuplicateValue = () => {
+Default.args = args
+
+export const WithDuplicateValue = args => {
   return (
     <Grid>
       <GridCol offset="1" col="8">
         <DropdownSelect
-          id={text('id', 'dropdown-select')}
-          error={boolean('error', false)}
-          valid={boolean('valid', false)}
-          disabled={boolean('disabled', false)}
-          hideLabel={boolean('hide label?', false)}
-          combobox={true}
-          labelText={text('LabelText', 'label')}
+          {...args}
+          id="dropdown-select-WithDuplicateValue"
+          uncontrolled
           options={[
             { value: 'france', label: 'France' },
             { value: 'france', label: 'France' },
             { value: 'irlande', label: 'Irlande' },
           ]}
-          size={select(
-            'size',
-            ['tiny', 'normal', 'big', 'huge', 'giant'],
-            'normal',
-          )}
-          variant={select('variant', ['andromeda', 'orion'], 'andromeda')}
           defaultSelectedValue="focus"
-          comboboxButtonLabelText={text('Buton aria-label', 'label')}
-          noResultText={text('No results text', 'No results')}
-          uniqLabelOnSearch={boolean('uniqLabelOnSearch', false)}
-          menuZIndex={number('menuZIndex', 1000)}
         />
       </GridCol>
     </Grid>
   )
 }
 
-export const WithComponentsForLabel = () => {
+WithDuplicateValue.args = args
+
+export const WithComponentsForLabel = args => {
   return (
     <Grid>
       <GridCol offset="1" col="8">
         <DropdownSelect
-          id={text('id', 'dropdown-select')}
-          error={boolean('error', false)}
-          valid={boolean('valid', false)}
-          disabled={boolean('disabled', false)}
-          hideLabel={boolean('hide label?', false)}
-          combobox={true}
-          labelText={text('LabelText', 'label')}
+          {...args}
+          id="dropdown-select-WithComponentsForLabel"
+          uncontrolled
           options={[
             {
               value: 'France',
@@ -164,24 +151,16 @@ export const WithComponentsForLabel = () => {
               ),
             },
           ]}
-          size={select(
-            'size',
-            ['tiny', 'normal', 'big', 'huge', 'giant'],
-            'normal',
-          )}
-          variant={select('variant', ['andromeda', 'orion'], 'andromeda')}
           defaultSelectedValue="focus"
-          comboboxButtonLabelText={text('Buton aria-label', 'label')}
-          noResultText={text('No results text', 'No results')}
-          uniqLabelOnSearch={boolean('uniqLabelOnSearch', false)}
-          menuZIndex={number('menuZIndex', 1000)}
         />
       </GridCol>
     </Grid>
   )
 }
 
-export const ControlledInput = () => {
+WithComponentsForLabel.args = args
+
+export const ControlledInput = args => {
   const [value, setValue] = useState('felt')
 
   const getRandomOption = () => {
@@ -199,7 +178,8 @@ export const ControlledInput = () => {
     <>
       <DropdownSelect
         id="ControlledInput"
-        combobox={true}
+        {...args}
+        labelText="Controlled Input"
         options={options}
         value={value}
         defaultSelectedValue="focus"
@@ -217,6 +197,11 @@ export const ControlledInput = () => {
       >
         Change Value
       </Button>
+      <Button className="k-u-margin-top-single" onClick={() => setValue('')}>
+        Reset Value
+      </Button>
     </>
   )
 }
+
+ControlledInput.args = { ...args, controlled: true }